refactored auth logic into home component
